requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
apsw-wheels==3.36.0.post1
    # via -r requirements.in
boto3==1.18.54
    # via
    #   -r requirements.in
    #   sqlite-s3vfs
botocore==1.21.54
    # via
    #   boto3
    #   s3transfer
click==8.0.1
    # via flask
flask==2.0.1
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
itsdangerous==2.0.1
    # via flask
jinja2==3.0.1
    # via flask
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
markupsafe==2.0.1
    # via jinja2
python-dateutil==2.8.2
    # via botocore
s3transfer==0.5.0
    # via boto3
six==1.16.0
    # via python-dateutil
sqlite-s3vfs==0.0.20
    # via -r requirements.in
urllib3==1.26.7
    # via botocore
werkzeug==2.0.1
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
